third_party/blink/web_tests/media/media-fragments/TC0081-expected.txt


Title: TC0081
Fragment: 't=0:00:00.'
Comment: A valid media fragment: { starting at 0 seconds. Thus, the UA will play the entire media resource.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK
END OF TEST
